Se añade el uso de JWT para la authenticacion en ciertos endpoints y el endpoint de cambio de dinero por el tipo de cambio que obtenga en base

+++ MoneyExchange.Application.DTO/ReceiveExchangeDto.cs
@@ -2,7 +2,7 @@
 {
     public class ReceiveExchangeDto
     {
-        public decimal Monto { get; set; }
+        public double Monto { get; set; }
         public string MonedaOrigen { get; set; }
         public string MonedaDestino { get; set; }
     }
